Task 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,4 @@
-#include "function_pointers.c"
+#include "function_pointers.h"
 #include <stdlib.h>
 
 /**
@@ -10,7 +10,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if(array == NULL || action == NULL)
 		return;
